Unified NSButton+Action and UIButton+Action
